Restructure main application and mixfile according to Elixir 1.4+ standards.
* Restore vanilla .gitignore
* Keep `mix.lock` versioned
* Move app initialization to NewRelixir.Application
* Reformat mix.exs
* Declare Ecto and Phoenix deps as `runtime: false`
* Declare ExDoc dep as `only: :docs`
